Moved parsing of a specific type into seperate functions.
The parseArgs function's argument type definition where used similar to
an enum and the parsing logic of a type was contained in parseArgs.

Now an argument type is a function containig the logic itself.
This makes parseArgs easier to read, a type's logic easier to find
and makes it easier to add new types.