Added switch to ignore stdout and stderr, use 'auto' mode for unittests
Fixed #118 by ignoring the output of stdout and stderr for tests
which write massive amounts of data to these streams. Since they
are buffered, the process deadlocks while waiting for those buffers
to be flushed, which actually never happens. The new class variables
`SUBPROCESS_CATCH_STDERR` and `SUBPROCESS_CATCH_STDOUR` allow a
subclass of the class `BaseTest` to circumvent this issue.